+++ feature/account/src/main/java/com/github/saikcaskey/account/data/datastore/AccountSettingsDataStoreImpl.kt
@@ -8,7 +8,7 @@ import androidx.datastore.preferences.core.emptyPreferences
 import androidx.datastore.preferences.core.intPreferencesKey
 import androidx.datastore.preferences.core.stringPreferencesKey
 import com.github.saikcaskey.pokertracker.domain.CoroutineDispatchers
-import com.github.saikcaskey.account.domain.datastore.AccountSettingsDataStore
+import com.github.saikcaskey.pokertracker.domain.datastore.AccountSettingsDataStore
 import com.github.saikcaskey.pokertracker.domain.models.AppInfo
 import com.github.saikcaskey.pokertracker.domain.models.AccountSettingsData
 import com.github.saikcaskey.pokertracker.domain.models.UserPreference
