+++ feature/account/src/main/java/com/github/saikcaskey/account/di/AccountSettingsDataSourceModule.kt
@@ -1,7 +1,7 @@
 package com.github.saikcaskey.account.di
 
 import com.github.saikcaskey.account.data.datasource.AccountSettingsDataSourceImpl
-import com.github.saikcaskey.account.domain.datasource.AccountSettingsDataSource
+import com.github.saikcaskey.pokertracker.domain.datasource.AccountSettingsDataSource
 import org.koin.dsl.module
 
 val accountSettingsDataSourceModule = module {
