+++ feature/account/src/main/java/com/github/saikcaskey/account/di/AccountSettingsDataStoreModule.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
 import androidx.datastore.preferences.preferencesDataStore
-import com.github.saikcaskey.account.domain.datastore.AccountSettingsDataStore
+import com.github.saikcaskey.pokertracker.domain.datastore.AccountSettingsDataStore
 import com.github.saikcaskey.account.data.datastore.AccountSettingsDataStoreImpl
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
